// frontend/src/features/financeiro/types/financeiro.types.ts
export interface FinanceiroResumo {
  totalGasto: number
  ganhoTotal: number
  lucroTotal: number
  receitasRealizadas: number
  receitasAReceber: number
  despesasRealizadas: number
  despesasAPagar: number
  despesasVencidas: number
  saldoRealizado: number
  saldoProjetado: number
  totalAnimaisCadastrados: number
  totalAnimaisAtivos: number
  totalAnimaisVendidos: number
  totalAnimaisInativos: number
  totalPastosCadastrados: number
  totalVendasRealizadas: number
}